+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Update minimum supported Rust version to 1.84.
 - Change Chiplet Fields to Public (#1629).
 - Added to the `Assembler` the ability to vendor a compiled library.
+- [BREAKING] Introduced `HORNERBASE`, `HORNEREXT` and removed `RCOMBBASE` (#1656).
 
 
 ## 0.12.0 (2025-01-22)